NUM-167 Moved AkkaMessageConsumer from provider to akka
