+++ amqp-driver/src/test/scala/com/nummulus/amqp/driver/DefaultAkkaConsumerTest.scala
@@ -7,11 +7,11 @@ import org.scalatest.FlatSpecLike
 import org.scalatest.Matchers
 import org.scalatest.junit.JUnitRunner
 
+import com.nummulus.amqp.driver.akka.AkkaMessageConsumer
 import com.nummulus.amqp.driver.api.consumer.AmqpConsumerResponse
 import com.nummulus.amqp.driver.api.consumer.AmqpConsumerRequest
 import com.nummulus.amqp.driver.api.provider.AmqpRequestMessageWithProperties
 import com.nummulus.amqp.driver.fixture.AkkaConsumerFixture
-import com.nummulus.amqp.driver.provider.AkkaMessageConsumer
 
 import _root_.akka.actor.ActorSystem
 import _root_.akka.testkit.ImplicitSender
