+++ amqp-driver/src/test/scala/com/nummulus/amqp/driver/akka/AkkaRabbitConsumerTest.scala
@@ -1,4 +1,4 @@
-package com.nummulus.amqp.driver.provider
+package com.nummulus.amqp.driver.akka
 
 import org.junit.runner.RunWith
 import org.scalatest.FlatSpecLike
